Implement student profile creation and listing
This commit introduces the ability to create new student profiles
and display a list of existing profiles in the admin panel. A new
`StudentProfile` model is used instead of directly using the `User`
model in the index view.

The index view now displays a list of `StudentProfile` records,
allowing admins to search and filter profiles. It displays student
information, including ID, GPA, and school name.